Assignment/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-cors-headers==3.6.0
django-webpush==0.3.3
djangorestframework==3.12.1
http-ece==1.1.0
idna==2.10
psycopg2==2.8.6
py-vapid==1.8.2
pycparser==2.20
pytz==2021.1
pywebpush==1.9.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
